SW-1541 Add complex objects support to Select
