+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@terraware/web-components",
-  "version": "1.0.11",
+  "version": "1.0.12",
   "author": "Terraformation Inc.",
   "license": "Apache-2.0",
   "repository": {
