+++ src/index.ts
@@ -20,6 +20,7 @@ import Note from './components/Note';
 import ProgressCircle from './components/ProgressCircle/ProgressCircle';
 import RadioButton from './components/RadioButton';
 import Select from './components/Select/Select';
+import SelectT from './components/Select/SelectT';
 import { Size } from './components/Size';
 import SummaryBox from './components/SummaryBox';
 import Textfield from './components/Textfield/Textfield';
@@ -58,6 +59,7 @@ export {
   ProgressCircle,
   RadioButton,
   Select,
+  SelectT,
   stableSort,
   SummaryBox,
   Svg,
